backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
sqlalchemy==2.0.43
pydantic-settings==2.11.0
PyMuPDF==1.26.4
python-docx==1.2.0
google-genai==1.29.0
PyMySQL==1.1.1
